events: handle registration deadline expired

+++ pages/events/[eventId]/index.tsx
@@ -13,6 +13,7 @@ import { EventRegistrationStatus } from '@/types/enums/event-registration-status
 import { EventRegistration } from '@/types/event-registration';
 import Link from 'next/link';
 import { routeLogin } from '@/types/consts/routes';
+import moment from 'moment';
 
 const Event = (props: any) => {
   const session = props.session;
@@ -47,27 +48,31 @@ const Event = (props: any) => {
       return;
     }
 
-    let url: string = `${process.env.NEXT_PUBLIC_BACKEND_URL}/v1/events/registration`;
-    let method: string = 'POST';
-    if (isRegistered()) {
-      url = `${process.env.NEXT_PUBLIC_BACKEND_URL}/v1/events/registration`;
-      method = 'DELETE';
-    }
+    if (event && event?.registrationDeadline > moment().unix()) {
+      let url: string = `${process.env.NEXT_PUBLIC_BACKEND_URL}/v1/events/registration`;
+      let method: string = 'POST';
+      if (isRegistered()) {
+        url = `${process.env.NEXT_PUBLIC_BACKEND_URL}/v1/events/registration`;
+        method = 'DELETE';
+      }
 
-    const response = await fetch(url, {
-      method: method,
-      body: JSON.stringify({
-        eventId: `${eventId}`,
-        username: `${session.user.username}`,
-      }),
-      headers: {
-        'Content-Type': 'application/json',
-        Authorization: `Bearer ${session.user.accessToken}`,
-      },
-    });
+      const response = await fetch(url, {
+        method: method,
+        body: JSON.stringify({
+          eventId: `${eventId}`,
+          username: `${session.user.username}`,
+        }),
+        headers: {
+          'Content-Type': 'application/json',
+          Authorization: `Bearer ${session.user.accessToken}`,
+        },
+      });
 
-    if (response.status == 200 || response.status == 201) {
-      router.back();
+      if (response.status == 200 || response.status == 201) {
+        router.reload();
+      }
+    } else {
+      console.log('Registration deadline exceeded.');
     }
   };
 
@@ -183,9 +188,11 @@ const Event = (props: any) => {
             <ActionButton action={Action.COPY} onClick={handleShareClicked} />
           </div>
 
-          <div className="ml-1">
-            <TextButton text={isRegistered() ? 'Unregister' : 'Register'} onClick={handleRegistrationClicked} />
-          </div>
+          {event.registrationDeadline > moment().unix() && (
+            <div className="ml-1">
+              <TextButton text={isRegistered() ? 'Unregister' : 'Register'} onClick={handleRegistrationClicked} />
+            </div>
+          )}
         </div>
       </div>
     </>
